requirements.txt
streamlit
groq
python-dotenv
ddgs
pandas
yfinance